third_party/blink/web_tests/http/tests/security/no-popup-from-sandbox-expected.txt

CONSOLE ERROR: Blocked opening 'about:blank' in a new window because the request was made in a sandboxed frame whose 'allow-popups' permission is not set.
CONSOLE MESSAGE: PASS
To run this test outside of DumpRenderTree, please disable your popup blocker!
